Extract fetch api to hook
